The function Get-UniqueDriveToFrom working
This function now will parse out all the drive letters from the supplied array of Hashtables/Objects and return a list of what the job will do such as "C, D, -> E" or "E -> C, D" for example.  This is only being displayed using Write-Verbose for now, but it is working and is ready to be implemented.  This is that should need be done is to display it to the user using Write-Host and formatting it to look good.