Modified Template (save as a csv).xlsx and AuotBackup.ps1 files
Now the file Template (save as a csv).xlsx correctly explains what to do to select specific files and also change the column label from FileExtension, which is not a accurate describer, to FileMatching.  Thus the required the every instance of FileExtension to be replaced with FileMatching inside the file AutoBackup.ps1.