Re-implemented and modified Copy-ItemWithProgress
I re-implemented and modified the above function because I thought it would fix an issue with a specific file being copied, however it doesn't work.  Also, I changed how the progress is displayed to the user so displays the Description field of the file in the progress.

There are two things that need to be modified before this script might be considered done.
1.  Need some logic inside the function "Copy-ItemWithProgress" that will determine if the file in "From" is different than the one in the "To" variable.  See the .NOTES section for more detail.
2.  Need some logic in the function "Start-Backup" either before (I think this would be best) or after the foreach loop that uses the function "CopyItemWithProgress" to check to see if there are extra files in the Destination folder that don't exist in the Source folder.  And if they don't exist in the Source, then you need to delete those files from the Destination.