Mostly finished implemtation
This script I would now say is 99.9% implemented.  It is now capable of checking individual items to see if they have been modified, in size or last write time, and it deletes files that are in the destination but are not in the source.

There is only 2 things I could add to this.
1.  See about moving the variables used for the function Assert-Progress into the function Start-Backup so that we will only have one copy of the data that we will need to modify.  Then inside the functions Get-RobocopyProgress and Copy-ItemWithProgress we will only have to modify the contents of the variables inside Start-Backup to change its information.
2.  Need to see about implementing Start-Threadjob for the counting of the files before the backup progress to speed up the counting of the files.

Other than these two things, this script is 100% functional.