Fixed Error #8
This was achieved by replacing "Compare-Object" cmdlet with "Test-Path" cmdlet to check to see if the file exists in the other folder when it is counting the files for the job.
This in theory should also speed up the counting process because we are removing an extra step from the process.