Remaster of the codebase
This remaster makes it so it is easier to follow along what the script is doing (like consistant naming sceme for objects/hashtables) and it runs faster by grabbing the data (such as the number of threads on the system) in a more efficient way.
I also moved all UI prints into a seperate file inside as series of functions so that it cleans up the main script file.
Next, I split up long functions into one main function and several sub-functions so they are easier to read.

This remaster also includes a few additions:
1.  You can now define network shares for the jobs.  In addition, the job's Source and Destiantion fields are validated while the script is reading them.  If any issues come up when reading them, such as a source/desination field is not supplied or a incorrectly formatted source/destination, then it will display these errors and will ask the user to fix this issues before it will allow you to continue.
2.  The user can continue with a backup/restore even if the source doesn't exist.  It will simply skip the entry when it is doing the backup/restore (Issue # 6 inside the doc file).