Added support for file/directory exclusion
This modification had added additional support for this script by making it so that it can exclude specific files/directories when backing up/restoring your files.  Read the "README.MD" file for some examples of how to use this new logic.

IMPORTANT: Because of this update any existing files that were created previously need to be updated by adding two additional columns to the files "ExcludedFiles" and "ExcludedDirectories" and any files that have something specified for the column "FileMatching" has to add "*" to the "ExcludedDirectories" column becuase how this script handles individual files have been modified.  Now by default, it is any file that matches and not just the files in the current directory.

Along side this set of changes, the script has also been made more user friendly by adding some additional colors for the UI elemnets.  I.E. headers are marked in dark yellow, errors are marked in red, etc.

Also, the counting of the files has been  reworked so that only files that have changed are counted instead of all the files that match the conditions provided.